models/__init__.py
"""Model backends and discovery utilities."""

from vtbench.models._base import ModelBackend
from vtbench.models._discover import discover_backend, list_backends

__all__ = ["ModelBackend", "discover_backend", "list_backends"]
